refactor: share one TLS handshake parser between HTTPS and QUIC DPI
Extract dpi/tls_common.rs from the duplicated parser stacks in https.rs and
quic/tls.rs. Transport differences (truncation policy, legacy version use,
partial ALPN, TLS 1.3 assumption) are captured in TlsParseOptions.

Deliberate behavior changes:
- TCP SNI now uses the strict hostname validator
- TCP handshake length cap raised from 16 KB to 64 KB
- the 100-extension cap now applies to QUIC too
