+++ crates/rustnet-core/src/network/dpi/quic/mod.rs
@@ -7,4 +7,4 @@ mod tls;
 mod tests;
 
 pub use packet::{is_quic_packet, parse_quic_packet};
-pub use tls::{is_partial_sni, try_extract_tls_from_reassembler};
+pub use tls::try_extract_tls_from_reassembler;
