+++ crates/rustnet-core/src/network/dpi/quic/tls.rs
@@ -1,158 +1,10 @@
-use crate::network::types::{CryptoFrameReassembler, TlsInfo, TlsVersion};
+use crate::network::dpi::tls_common::{
+    self, TlsParseOptions, is_partial_sni, parse_sni_extension, parse_sni_header,
+};
+use crate::network::types::{CryptoFrameReassembler, TlsInfo};
 use log::debug;
 
 use super::salvage::{try_parse_unencrypted_crypto_frames, try_reconstruct_sni_from_fragments};
-
-// ============================================================================
-// SNI Validation and Parsing Helpers
-// ============================================================================
-
-/// Minimum length for partial SNI extraction
-const PARTIAL_SNI_MIN_LENGTH: usize = 3;
-
-/// Marker suffix for partial SNI values
-const PARTIAL_SNI_MARKER: &str = "[PARTIAL]";
-
-/// Validate if a string looks like a valid complete hostname
-///
-/// This is the unified hostname validation function used across all SNI extraction methods.
-/// Rules:
-/// - Length between 4 and 253 characters
-/// - Contains at least one '.'
-/// - Only ASCII alphanumeric, '.', and '-' characters
-/// - Doesn't start or end with '.' or '-'
-/// - No consecutive dots '..'
-/// - Has at least one alphabetic character
-/// - Each label is non-empty and at most 63 characters
-pub(super) fn is_valid_hostname(hostname: &str) -> bool {
-    // Length check
-    if hostname.len() < 4 || hostname.len() > 253 {
-        return false;
-    }
-
-    // Must contain at least one dot
-    if !hostname.contains('.') {
-        return false;
-    }
-
-    // Check for valid hostname characters only
-    if !hostname
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || c == '.' || c == '-')
-    {
-        return false;
-    }
-
-    // Must not start or end with a dot or hyphen
-    if hostname.starts_with('.')
-        || hostname.ends_with('.')
-        || hostname.starts_with('-')
-        || hostname.ends_with('-')
-    {
-        return false;
-    }
-
-    // Must not contain consecutive dots
-    if hostname.contains("..") {
-        return false;
-    }
-
-    // Must have at least one alphabetic character (not just numbers and dots)
-    if !hostname.chars().any(|c| c.is_ascii_alphabetic()) {
-        return false;
-    }
-
-    // Each label must be non-empty and at most 63 characters
-    if !hostname
-        .split('.')
-        .all(|part| !part.is_empty() && part.len() <= 63)
-    {
-        return false;
-    }
-
-    true
-}
-
-/// Validate if a string looks like a valid partial hostname
-///
-/// Partial hostnames have relaxed rules since they may be truncated:
-/// - Minimum length (PARTIAL_SNI_MIN_LENGTH)
-/// - Only ASCII alphanumeric, '.', and '-' characters
-/// - Has at least one alphabetic character
-/// - Doesn't start with '.' or '-'
-fn is_valid_partial_hostname(hostname: &str) -> bool {
-    if hostname.len() < PARTIAL_SNI_MIN_LENGTH {
-        return false;
-    }
-
-    // Check for valid hostname characters only
-    if !hostname
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || c == '.' || c == '-')
-    {
-        return false;
-    }
-
-    // Must have at least one alphabetic character
-    if !hostname.chars().any(|c| c.is_ascii_alphabetic()) {
-        return false;
-    }
-
-    // Must not start with '.' or '-'
-    if hostname.starts_with('.') || hostname.starts_with('-') {
-        return false;
-    }
-
-    true
-}
-
-/// Mark an SNI value as partial by appending the marker
-fn mark_partial_sni(hostname: &str) -> String {
-    format!("{}{}", hostname, PARTIAL_SNI_MARKER)
-}
-
-/// Check if an SNI value is marked as partial
-pub fn is_partial_sni(sni: &str) -> bool {
-    sni.ends_with(PARTIAL_SNI_MARKER)
-}
-
-/// Parsed SNI extension header
-struct SniHeader {
-    /// Server name list length
-    list_len: u16,
-    /// Hostname length
-    name_len: u16,
-}
-
-/// Parse the SNI extension header from raw data
-///
-/// Expects data starting at the SNI extension content (after extension type and length):
-/// - 2 bytes: server name list length
-/// - 1 byte: name type (0x00 = hostname)
-/// - 2 bytes: hostname length
-///
-/// Returns None if data is too short or name type is not hostname
-fn parse_sni_header(data: &[u8]) -> Option<SniHeader> {
-    if data.len() < 5 {
-        return None;
-    }
-
-    let list_len = u16::from_be_bytes([data[0], data[1]]);
-    let name_type = data[2];
-    let name_len = u16::from_be_bytes([data[3], data[4]]);
-
-    // Name type must be 0x00 (hostname)
-    if name_type != 0x00 {
-        return None;
-    }
-
-    // Validate hostname length is reasonable
-    if name_len == 0 || name_len > 253 {
-        return None;
-    }
-
-    Some(SniHeader { list_len, name_len })
-}
 
 /// Check if SNI is complete (not partial)
 fn is_complete_sni(sni: &Option<String>) -> bool {
@@ -359,54 +211,9 @@ pub fn try_extract_tls_from_reassembler(
 
 /// Parse a TLS handshake from reassembled data
 pub(super) fn parse_partial_tls_handshake(data: &[u8], allow_partial: bool) -> Option<TlsInfo> {
-    if data.len() < 4 {
-        debug!("QUIC: TLS handshake data too short: {} bytes", data.len());
-        return None;
-    }
-
-    let handshake_type = data[0];
-    let handshake_length = u32::from_be_bytes([0, data[1], data[2], data[3]]) as usize;
-
-    debug!(
-        "QUIC: TLS handshake type=0x{:02x}, declared_length={}, available_data={}, allow_partial={}",
-        handshake_type,
-        handshake_length,
-        data.len() - 4,
-        allow_partial
-    );
-
     let mut info = TlsInfo::new();
-
-    let available_data = &data[4..];
-    let parse_length = handshake_length.min(available_data.len());
-
-    // Sanity check the handshake length
-    if handshake_length > 65536 {
-        debug!(
-            "QUIC: Handshake length {} seems too large, skipping",
-            handshake_length
-        );
+    if !tls_common::parse_handshake(data, &mut info, TlsParseOptions::quic(allow_partial)) {
         return None;
-    }
-
-    match handshake_type {
-        0x01 => {
-            // Client Hello
-            debug!("QUIC: Parsing ClientHello with {} bytes", parse_length);
-            parse_partial_client_hello(&available_data[..parse_length], &mut info, allow_partial);
-        }
-        0x02 => {
-            // Server Hello
-            debug!("QUIC: Parsing ServerHello with {} bytes", parse_length);
-            parse_partial_server_hello(&available_data[..parse_length], &mut info);
-        }
-        _ => {
-            debug!(
-                "QUIC: Unknown/unsupported handshake type: 0x{:02x}",
-                handshake_type
-            );
-            return None;
-        }
     }
 
     debug!(
@@ -419,332 +226,6 @@ pub(super) fn parse_partial_tls_handshake(data: &[u8], allow_partial: bool) -> O
     } else {
         debug!("QUIC: No useful TLS info extracted");
         None
-    }
-}
-
-/// Parse a partial Client Hello
-fn parse_partial_client_hello(data: &[u8], info: &mut TlsInfo, allow_partial: bool) {
-    debug!(
-        "QUIC: Parsing ClientHello with {} bytes (allow_partial={})",
-        data.len(),
-        allow_partial
-    );
-
-    if data.len() < 34 {
-        debug!(
-            "QUIC: ClientHello too short: {} bytes (need at least 34)",
-            data.len()
-        );
-        return;
-    }
-
-    // Skip version (2) + random (32)
-    let mut offset = 34;
-    debug!(
-        "QUIC: ClientHello - skipping version and random, offset now={}",
-        offset
-    );
-
-    // Session ID
-    if offset >= data.len() {
-        return;
-    }
-    let session_id_len = data[offset] as usize;
-    offset += 1 + session_id_len;
-
-    if offset + 2 > data.len() {
-        return;
-    }
-
-    // Cipher suites
-    let cipher_suites_len = u16::from_be_bytes([data[offset], data[offset + 1]]) as usize;
-    offset += 2 + cipher_suites_len;
-
-    if offset >= data.len() {
-        return;
-    }
-
-    // Compression methods
-    let compression_len = data[offset] as usize;
-    offset += 1 + compression_len;
-
-    if offset + 2 > data.len() {
-        return;
-    }
-
-    // Extensions
-    let extensions_len = u16::from_be_bytes([data[offset], data[offset + 1]]) as usize;
-    offset += 2;
-
-    debug!(
-        "QUIC: ClientHello extensions - declared_len={}, available_data={}",
-        extensions_len,
-        data.len() - offset
-    );
-
-    let available_ext_len = (data.len() - offset).min(extensions_len);
-    if available_ext_len > 0 {
-        debug!("QUIC: Parsing {} bytes of extensions", available_ext_len);
-        parse_tls_extensions(
-            &data[offset..offset + available_ext_len],
-            info,
-            true,
-            allow_partial,
-        );
-    } else {
-        debug!("QUIC: No extensions data available");
-    }
-}
-
-/// Parse a partial Server Hello
-fn parse_partial_server_hello(data: &[u8], info: &mut TlsInfo) {
-    if data.len() < 34 {
-        return;
-    }
-
-    // Skip version (2) + random (32)
-    let mut offset = 34;
-
-    // Session ID
-    if offset >= data.len() {
-        return;
-    }
-    let session_id_len = data[offset] as usize;
-    offset += 1 + session_id_len;
-
-    if offset + 2 > data.len() {
-        return;
-    }
-
-    // Cipher suite
-    let cipher = u16::from_be_bytes([data[offset], data[offset + 1]]);
-    info.cipher_suite = Some(cipher);
-    offset += 2;
-
-    // Compression method
-    if offset >= data.len() {
-        return;
-    }
-    offset += 1;
-
-    // Extensions
-    if offset + 2 > data.len() {
-        return;
-    }
-
-    let extensions_len = u16::from_be_bytes([data[offset], data[offset + 1]]) as usize;
-    offset += 2;
-
-    let available_ext_len = (data.len() - offset).min(extensions_len);
-    if available_ext_len > 0 {
-        parse_tls_extensions(
-            &data[offset..offset + available_ext_len],
-            info,
-            false,
-            false,
-        );
-    }
-}
-
-/// Parse TLS extensions
-fn parse_tls_extensions(data: &[u8], info: &mut TlsInfo, is_client: bool, allow_partial: bool) {
-    let mut offset = 0;
-    debug!(
-        "QUIC: Parsing {} bytes of TLS extensions (is_client={}, allow_partial={})",
-        data.len(),
-        is_client,
-        allow_partial
-    );
-
-    while offset + 4 <= data.len() {
-        let ext_type = u16::from_be_bytes([data[offset], data[offset + 1]]);
-        let ext_len = u16::from_be_bytes([data[offset + 2], data[offset + 3]]) as usize;
-
-        debug!(
-            "QUIC: Extension type=0x{:04x}, length={}",
-            ext_type, ext_len
-        );
-
-        if offset + 4 + ext_len > data.len() {
-            // Extension data is incomplete
-            if allow_partial && ext_type == 0x0000 && is_client {
-                // Try to extract partial SNI as fallback
-                let available_ext_len = data.len() - offset - 4;
-                if available_ext_len > 5 {
-                    debug!(
-                        "QUIC: SNI extension is incomplete (need {} bytes, have {}), attempting partial extraction",
-                        ext_len, available_ext_len
-                    );
-                    let ext_data = &data[offset + 4..];
-                    if let Some(sni) = parse_sni_extension(ext_data, true) {
-                        debug!("QUIC: Extracted partial SNI as fallback: {}", sni);
-                        info.sni = Some(sni);
-                    }
-                }
-            } else {
-                debug!(
-                    "QUIC: Extension 0x{:04x} is incomplete (need {} bytes, have {}), waiting for more data",
-                    ext_type,
-                    ext_len,
-                    data.len() - offset - 4
-                );
-            }
-            break;
-        }
-
-        let ext_data = &data[offset + 4..offset + 4 + ext_len];
-
-        match ext_type {
-            0x0000 if is_client => {
-                // SNI
-                debug!("QUIC: Found SNI extension with {} bytes", ext_len);
-                if let Some(sni) = parse_sni_extension(ext_data, allow_partial) {
-                    debug!("QUIC: Successfully parsed SNI: {}", sni);
-                    info.sni = Some(sni);
-                } else {
-                    debug!("QUIC: Failed to parse SNI extension");
-                }
-            }
-            0x0010 => {
-                // ALPN
-                debug!("QUIC: Found ALPN extension with {} bytes", ext_len);
-                if let Some(alpn) = parse_alpn_extension(ext_data) {
-                    debug!("QUIC: Successfully parsed ALPN: {:?}", alpn);
-                    info.alpn = alpn;
-                } else {
-                    debug!("QUIC: Failed to parse ALPN extension");
-                }
-            }
-            0x002b => {
-                // Supported Versions
-                debug!(
-                    "QUIC: Found Supported Versions extension with {} bytes",
-                    ext_len
-                );
-                if let Some(version) = parse_supported_versions(ext_data, is_client) {
-                    debug!("QUIC: Successfully parsed version: {:?}", version);
-                    info.version = Some(version);
-                }
-            }
-            _ => {
-                debug!("QUIC: Skipping unknown extension type 0x{:04x}", ext_type);
-            }
-        }
-
-        offset += 4 + ext_len;
-    }
-
-    debug!(
-        "QUIC: Finished parsing extensions - found SNI={:?}, ALPN={:?}",
-        info.sni, info.alpn
-    );
-}
-
-/// Parse SNI extension
-fn parse_sni_extension(data: &[u8], allow_partial: bool) -> Option<String> {
-    debug!(
-        "QUIC: Parsing SNI extension with {} bytes (allow_partial={}): {:02x?}",
-        data.len(),
-        allow_partial,
-        &data[..data.len().min(20)]
-    );
-
-    // Parse the SNI header using the unified helper
-    let header = match parse_sni_header(data) {
-        Some(h) => h,
-        None => {
-            debug!(
-                "QUIC: Failed to parse SNI header (data len: {})",
-                data.len()
-            );
-            return None;
-        }
-    };
-
-    debug!(
-        "QUIC: SNI header - list_len: {}, name_len: {}",
-        header.list_len, header.name_len
-    );
-
-    let name_len = header.name_len as usize;
-    let hostname_start = 5; // After header (2 + 1 + 2 bytes)
-
-    if hostname_start + name_len <= data.len() {
-        // Full hostname available
-        let sni_data = &data[hostname_start..hostname_start + name_len];
-        debug!("QUIC: SNI data: {:02x?}", sni_data);
-
-        match std::str::from_utf8(sni_data) {
-            Ok(sni) => {
-                if is_valid_hostname(sni) {
-                    debug!("QUIC: Successfully parsed complete SNI: {}", sni);
-                    Some(sni.to_string())
-                } else {
-                    debug!("QUIC: SNI doesn't look like a valid hostname: {}", sni);
-                    None
-                }
-            }
-            Err(e) => {
-                debug!("QUIC: SNI data is not valid UTF-8: {}", e);
-                None
-            }
-        }
-    } else if allow_partial && data.len() > hostname_start {
-        // Extract partial SNI as fallback when allowed
-        let available = &data[hostname_start..];
-        debug!(
-            "QUIC: SNI name extends beyond available data (need {}, have {}), extracting partial",
-            hostname_start + name_len,
-            data.len()
-        );
-
-        if let Ok(partial) = std::str::from_utf8(available)
-            && is_valid_partial_hostname(partial)
-        {
-            debug!("QUIC: Extracted partial SNI: {}", mark_partial_sni(partial));
-            return Some(mark_partial_sni(partial));
-        }
-        None
-    } else {
-        // SNI data is incomplete - don't extract partial, wait for more data
-        debug!(
-            "QUIC: SNI name extends beyond available data (need {}, have {}), waiting for more data",
-            hostname_start + name_len,
-            data.len()
-        );
-        None
-    }
-}
-
-/// Parse ALPN extension
-pub(super) fn parse_alpn_extension(data: &[u8]) -> Option<Vec<String>> {
-    if data.len() < 2 {
-        return None;
-    }
-
-    let mut protocols = Vec::new();
-    let alpn_len = u16::from_be_bytes([data[0], data[1]]) as usize;
-
-    let mut offset = 2;
-    let list_end = 2 + alpn_len.min(data.len() - 2);
-
-    while offset < list_end && offset < data.len() {
-        let proto_len = data[offset] as usize;
-        offset += 1;
-
-        if offset + proto_len <= data.len()
-            && let Ok(proto) = std::str::from_utf8(&data[offset..offset + proto_len])
-        {
-            protocols.push(proto.to_string());
-        }
-
-        offset += proto_len;
-    }
-
-    if protocols.is_empty() {
-        None
-    } else {
-        Some(protocols)
     }
 }
 
@@ -843,30 +324,6 @@ pub(super) fn scan_for_sni_extension(
     (0..data.len()).find_map(|i| match_sni_extension_at(data, i, allow_partial, strictness))
 }
 
-/// Parse supported versions extension
-fn parse_supported_versions(data: &[u8], is_client: bool) -> Option<TlsVersion> {
-    if is_client {
-        if data.is_empty() {
-            return None;
-        }
-
-        let list_len = data[0] as usize;
-        let mut offset = 1;
-
-        while offset + 1 < data.len() && offset < 1 + list_len {
-            if data[offset] == 0x03 && data[offset + 1] == 0x04 {
-                return Some(TlsVersion::Tls13);
-            }
-            offset += 2;
-        }
-    } else if data.len() >= 2 && data[0] == 0x03 && data[1] == 0x04 {
-        return Some(TlsVersion::Tls13);
-    }
-
-    // QUIC always uses TLS 1.3
-    Some(TlsVersion::Tls13)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -933,77 +390,6 @@ mod tests {
         // With allow_partial=true, returns partial SNI
         let result = scan_for_sni_extension(&data, true, SniScanStrictness::Lenient);
         assert_eq!(result, Some("www.examp[PARTIAL]".to_string()));
-    }
-
-    #[test]
-    fn test_parse_sni_extension_complete() {
-        // Build SNI extension data (without the extension type/length header)
-        let hostname = "test.example.org";
-        let name_bytes = hostname.as_bytes();
-        let name_len = name_bytes.len() as u16;
-        let list_len = name_len + 3;
-
-        let mut data = Vec::new();
-        data.extend_from_slice(&list_len.to_be_bytes());
-        data.push(0x00); // name type
-        data.extend_from_slice(&name_len.to_be_bytes());
-        data.extend_from_slice(name_bytes);
-
-        let result = parse_sni_extension(&data, false);
-        assert_eq!(result, Some("test.example.org".to_string()));
-    }
-
-    #[test]
-    fn test_parse_sni_extension_partial() {
-        // Build partial SNI extension data
-        let mut data = Vec::new();
-        data.extend_from_slice(&20u16.to_be_bytes()); // list_len
-        data.push(0x00); // name type
-        data.extend_from_slice(&15u16.to_be_bytes()); // declared name_len
-        data.extend_from_slice(b"example.co"); // only 10 chars
-
-        // With allow_partial=false, returns None
-        let result = parse_sni_extension(&data, false);
-        assert_eq!(result, None);
-
-        // With allow_partial=true, returns partial SNI
-        let result = parse_sni_extension(&data, true);
-        assert_eq!(result, Some("example.co[PARTIAL]".to_string()));
-    }
-
-    #[test]
-    fn test_parse_alpn_extension() {
-        // Build ALPN extension data (without the extension type/length header)
-        let mut data = Vec::new();
-        let protocols = vec!["h3", "h2"];
-
-        let mut proto_list = Vec::new();
-        for proto in &protocols {
-            proto_list.push(proto.len() as u8);
-            proto_list.extend_from_slice(proto.as_bytes());
-        }
-
-        data.extend_from_slice(&(proto_list.len() as u16).to_be_bytes());
-        data.extend_from_slice(&proto_list);
-
-        let result = parse_alpn_extension(&data);
-        assert_eq!(result, Some(vec!["h3".to_string(), "h2".to_string()]));
-    }
-
-    #[test]
-    fn test_is_valid_hostname() {
-        assert!(is_valid_hostname("example.com"));
-        assert!(is_valid_hostname("www.example.com"));
-        assert!(is_valid_hostname("sub.domain.example.org"));
-        assert!(is_valid_hostname("my-site.io"));
-
-        // Invalid hostnames
-        assert!(!is_valid_hostname("com")); // No dot
-        assert!(!is_valid_hostname(".example.com")); // Starts with dot
-        assert!(!is_valid_hostname("example.com.")); // Ends with dot
-        assert!(!is_valid_hostname("-example.com")); // Starts with hyphen
-        assert!(!is_valid_hostname("example..com")); // Consecutive dots
-        assert!(!is_valid_hostname("ab")); // Too short
     }
 
     #[test]
